add edit profile to testscren rootstack
