+++ navigators/RootStackNavigator.tsx
@@ -14,6 +14,7 @@ import TestMembers from '../screens/debug/TestMembers';
 import TestRequests from '../screens/debug/TestRequests';
 import TestTasks from '../screens/debug/TestTasks';
 import TestUser from '../screens/debug/TestUser';
+import EditProfileScreen from '../screens/EditProfileScreen';
 import EditTaskScreen from '../screens/EditTaskScreen';
 import HouseholdInformationScreen from '../screens/HouseholdInformationScreen';
 import JoinHouseholdScreen from '../screens/JoinHouseholdScreen';
@@ -51,6 +52,7 @@ export type RootStackParamList = {
   TestHouseholds: undefined;
   TestCompTasks: undefined;
   EditTask: { task: Task };
+  EditProfile: undefined;
   TestRequests: undefined;
 };
 
@@ -90,6 +92,7 @@ export default function RootStackNavigator() {
             component={TaskInfoScreen}
             options={{ title: 'Information om syssla' }}
           />
+          <RootStack.Screen name="EditProfile" component={EditProfileScreen} />
 
           <RootStack.Screen
             name="SelectedHouseholdNav"
